fix: use mux for forwarder, break out test package
This is a small refactor to use the handler mux in the forwarder and
break out the mock utilitiy into a separate package.
